Move UiiInfo to model Filter displayed Content types

+++ phapi/admin/Controllers/ContentManager.php
@@ -2,6 +2,7 @@
 
 namespace admin\Controllers;
 
+use Exception;
 use Phapi;
 use Phapi\Controller;
 
@@ -20,33 +21,51 @@ class ContentManager extends Controller {
 
     public function findOne($params, $next) {
         $model = Phapi::model($params["model"]);
+        $uiInfo = $model->getUiInfo();
+        if (!$uiInfo)
+            throw new Exception("Using this model is not allowed", 403);
         return $model->findOne([$model->getPrimaryField() => $params['id']]);
     }
 
     public function find($params) {
         $model = Phapi::model($params["model"]);
+        $uiInfo = $model->getUiInfo();
+        if (!$uiInfo)
+            throw new Exception("Using this model is not allowed", 403);
         return $model->find([]);
     }
 
     public function create($params) {
         $model = Phapi::model($params["model"]);
+        $uiInfo = $model->getUiInfo();
+        if (!$uiInfo || !$uiInfo["editable"])
+            throw new Exception("Using this model is not allowed", 403);
         $data = json_load_body();
         return $model->create($data);
     }
 
     public function update($params, $next) {
         $model = Phapi::model($params["model"]);
+        $uiInfo = $model->getUiInfo();
+        if (!$uiInfo || !$uiInfo["editable"])
+            throw new Exception("Using this model is not allowed", 403);
         $data = json_load_body();
         return $model->update([$model->getPrimaryField() => $params['id']], $data);
     }
 
     public function delete($params, $next) {
         $model = Phapi::model($params["model"]);
+        $uiInfo = $model->getUiInfo();
+        if (!$uiInfo || !$uiInfo["editable"])
+            throw new Exception("Using this model is not allowed", 403);
         return $model->delete([$model->getPrimaryField() => $params['id']]);
     }
 
     public function count($params) {
         $model = Phapi::model($params["model"]);
+        $uiInfo = $model->getUiInfo();
+        if (!$uiInfo)
+            throw new Exception("Using this model is not allowed", 403);
         return $model->count(null);
     }
 }
